requirements.txt
absl-py
flax~=0.8.5
jax>=0.4.28
jaxlib>=0.4.28
lightning
lightning-utilities
matplotlib~=3.9.0
ml_collections
optax~=0.2.2
orbax-checkpoint
penzai
pytorch-lightning
PyYAML~=6.0.1
tabulate~=0.9.0
tensorboard
tensorboardX
tensorflow
tensorflow-datasets
torch~=2.4.0.dev20240601+cu124
torchmetrics
torchvision~=0.19.0.dev20240601+cu124
tqdm~=4.66.4
wandb

setuptools~=69.5.1
numpy~=1.26.4
pillow~=9.3.0